requirements.txt
numpy==1.26.4
opencv_python_headless==4.10.0.84
pdf2image==1.17.0
Pillow==11.0.0
sahi
torch==2.0.1
ultralytics==8.1.34
huggingface-hub
